requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.35
transformers==4.36.2
torch==2.1.2